Split config into individual files.
